✨[Feat] block post api 호출하는 함수 생성 #111
