+++ src/app/admin/block/components/layout.tsx
@@ -2,6 +2,7 @@
 import React, { FormEvent } from "react";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
+import QuestionIcon from "@app/admin/block/components/question-icon";
 
 const Layout = ({
   title,
@@ -9,7 +10,7 @@ const Layout = ({
   children,
 }: Readonly<{
   title: string;
-  onSubmit: (e: FormEvent<HTMLFormElement>) => void;
+  onSubmit?: (e: FormEvent<HTMLFormElement>) => void;
   children: React.ReactNode;
 }>) => {
   const router = useRouter();
@@ -27,21 +28,7 @@ const Layout = ({
       </div>
       <div className="flex items-center gap-1">
         <h1 className="pageName">{title}</h1>
-        <div className="group relative inline-block">
-          <Image
-            src="/assets/icons/icon_help.png"
-            alt="question"
-            width={30}
-            height={30}
-          />
-          {title === "링크 블록" && (
-            <div className="absolute left-2 top-20 w-max -translate-x-1/2 -translate-y-1/2 transform rounded bg-[#343434] px-3 py-3 text-sm text-white opacity-0 transition-opacity duration-300 group-hover:opacity-100">
-              • 기본 정보와 공개 여부 값은 필수입니다. <br />• 예약 공개와
-              스티커는 프로 기능입니다.
-              <div className="absolute -top-3 left-1/2 h-0 w-0 rotate-90 border-y-8 border-r-8 border-y-transparent border-r-[#343434]"></div>
-            </div>
-          )}
-        </div>
+        <QuestionIcon title={title} />
       </div>
       <form onSubmit={onSubmit} className="flex flex-col gap-4">
         {children}
